test: extract bigint test a isolated file
