+++ packages/base/tests/utils.test.js
@@ -5,12 +5,6 @@ const { JSBI } = require("../lib/primitive");
 const {
   CKBHasher,
   ckbHash,
-
-  toBigUInt64LE,
-  readBigUInt64LE,
-  readBigUInt128LE,
-  toBigUInt128LE,
-
   toBigUInt64LECompatible,
   readBigUInt64LECompatible,
   readBigUInt128LECompatible,
@@ -46,38 +40,6 @@ test("ckbHash", (t) => {
   const arrayBuffer = new Reader(message).toArrayBuffer();
   const result = ckbHash(arrayBuffer);
   t.is(result.serializeJson(), messageDigest);
-});
-
-const uint64 = 1965338n;
-const uint64le = "0x1afd1d0000000000";
-
-test("BigInt:toBigUInt64LE", (t) => {
-  t.is(toBigUInt64LE(uint64), uint64le);
-});
-
-test("BigInt:readBigUInt64LE", (t) => {
-  t.is(readBigUInt64LE(uint64le), uint64);
-});
-
-const u128 = BigInt("1208925819614629174706177");
-const u128le = "0x01000000000000000000010000000000";
-
-test("BigInt:toBigUInt128LE", (t) => {
-  t.is(toBigUInt128LE(u128), u128le);
-});
-
-test("BigInt:toBigUInt128LE, to small", (t) => {
-  t.throws(() => toBigUInt128LE(-1n));
-  t.notThrows(() => toBigUInt128LE(0n));
-});
-
-test("BigInt:toBigUInt128LE, to big", (t) => {
-  t.throws(() => toBigUInt128LE(2n ** 128n));
-  t.notThrows(() => toBigUInt128LE(2n ** 128n - 1n));
-});
-
-test("BigInt:readBigUInt128LE", (t) => {
-  t.is(readBigUInt128LE(u128le), u128);
 });
 
 const uint64Compatible = JSBI.BigInt(1965338);
